Fix build: move carry-over extension into a compiled file
The Stat module uses explicit Xcode file references (not a synchronized
group), so the standalone RollingAverageCarryOver.swift was never compiled,
causing 'no member updateMealCarryOver / tddCarryOverValue' archive errors.
Move the carry-over extension into StatStateModel.swift, which is already
referenced by the project.
